// src/types/ProductFormValues.ts
export type ProductFormValues = {
  name: string;
  vendor: string;
  brand: string;
  url: string;
  sku: string;
  color: string;
  size: string;
  category: string;
  family: string;
  price: number;
  size_chart_link: string;
  how_to_measure_link: string;
  decoration_method: string;
  setup_cost: number;
  setup_cost_code: string;
  deco_cost: number;
  moq: number;
  production_time: string;
  shipping_weight: number;
}